Further improvements for small screens
Reduced the margin at the top and bottom.  It's a bit fiddly - both the top margin and the height calculation need to be done together or the bottom shifts out of view.

Reduced the box-shadow which adds a subtle page edge, which doesn't need to be as big on a small screen.